Added AI to main.py
Created AI players that will either utilize a sequential list of moves to decide their turn or a random list of moves to make a choice. which AI is against which AI can be decided in lines 243 and 245 NOTE: placing two sequential AI against each other will always result in a cat's game.